marketplace/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-autoslug==1.9.8
psycopg2-binary==2.9.3
pytils==0.4.1
sqlparse==0.4.2
tzdata==2022.1
